requirements.txt
fastapi 
uvicorn 
crewai 
langchain-groq
langsmith
pygithub 
python-dotenv
langgraph